Move passport-config to util
