+++ routes/index.js
@@ -8,7 +8,7 @@ const passport = require("passport");
 const flash = require("express-flash");
 const session = require("express-session");
 
-const passport_config = require("../passport-config");
+const passport_config = require("../util/passport-config");
 
 const initializePassport = passport_config.initialize;
 const checkAuthenticated = passport_config.checkAuthenticated;
